requirements.txt
# Core dependencies
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# LLM providers
anthropic>=0.39.0
openai>=1.12.0
google-generativeai>=0.8.0

# Web search
tavily-python>=0.3.0

# Web scraping and content extraction
httpx>=0.27.0
trafilatura>=1.6.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Vector database
chromadb>=0.4.22

# CLI and UI
rich>=13.7.0
typer>=0.9.0

# Async support
asyncio>=3.4.3
aiohttp>=3.9.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# Development
black>=24.0.0
ruff>=0.1.0
mypy>=1.8.0
